db-sync: Config refactoring to allow generation of LedgerState config
In order to correctly build and maintain ledger state 'db-sync'
needs to construct a 'LedgerState' configuration. This was incredibly
difficult because the ledger state config is buried within a
'TopLevelConfig' and building that is difficult because it requires
a bunch of config parameters that 'db-sync' does not otherwise care
about.